+++ CHANGELOG.md
@@ -43,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `LexiconCollection` and `MWELexiconCollection` open the TSV file downloaded through `from_tsv` method by default using `utf-8` encoding.
 - `pymusas_rule_based_tagger` is now a spacy registered factory by using an entry point.
 - `MWELexiconCollection` warns users that it does not support curly braces MWE template expressions.
+- All of the POS mappings can now be called through a spaCy registered function, all of these functions can be found in the `pymusas.spacy_api.pos_mapper` module.
 
 ### Removed